Revert "[android] mme renderer now works on Android"
This reverts commit 051999a2fe2e524af7c3ede732498bd7b687db17.

+++ codemp/rd-vanilla/tr_surfacesprites.cpp
@@ -1407,7 +1407,7 @@ extern void R_InvertMatrix(float *sourcemat, float *destmat);
 
 void RB_DrawSurfaceSprites( shaderStage_t *stage, shaderCommands_t *input) 
 {
-	uint32_t	glbits=stage->stateBits;
+	unsigned long	glbits=stage->stateBits;
 	
 	R_SurfaceSpriteFrameUpdate();
